Add plugin support
